requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
yt-dlp==2024.7.9
youtube-search-python==1.6.6
python-dotenv==1.0.0
validators==0.33.0g